+++ resources/views/Adotado/show.blade.php
@@ -9,27 +9,15 @@
             <div class="col-8">
                 <h3 class="card-title">ID: {{$adotado->id}}</h3> <br>
             </div>
+            @if ((Auth::check()) && (Auth::user()->isAdmin()))
                 <div class="col-4">
-                    @if ($adotado->data_visita_um == null)    
-                        {{Form::open(['route'=>['adotados.visitarum', $adotado->id], 'method'=>'PUT'])}}
+                    @if ($adotado->datavisita == null)    
+                        {{Form::open(['route'=>['adotados.visitar', $adotado->id], 'method'=>'PUT'])}}
                         {{Form::submit('Visitar', ['class'=>'btn btn-outline-success', 'onclick'=>'return confim("Confirma visita?")'])}}
                         {{Form::close()}}
                     @endif
                 </div>
-                <div class="col-4">
-                    @if ($adotado->data_visita_dois == null)    
-                        {{Form::open(['route'=>['adotados.visitardois', $adotado->id], 'method'=>'PUT'])}}
-                        {{Form::submit('Visitar', ['class'=>'btn btn-outline-success', 'onclick'=>'return confim("Confirma visita?")'])}}
-                        {{Form::close()}}
-                    @endif
-                </div>
-                <div class="col-4">
-                    @if ($adotado->data_visita_tres == null)    
-                        {{Form::open(['route'=>['adotados.visitartres', $adotado->id], 'method'=>'PUT'])}}
-                        {{Form::submit('Visitar', ['class'=>'btn btn-outline-success', 'onclick'=>'return confim("Confirma visita?")'])}}
-                        {{Form::close()}}
-                    @endif
-                </div>
+            @endif    
         </div>
     </div>
     <div class="card-body">
@@ -38,16 +26,19 @@
             <strong>Adotante:</strong> {{$adotado->adotante_id}} - {{$adotado->adotante->nome}}<br>
             <strong>Data:</strong> {{\Carbon\Carbon::create($adotado->datahora)->format('d/mY H:i:s')}}<br>
             <strong>Obs:</strong> {{$adotado->obs}}<br>
-            <strong>Data 1º visita:</strong> {{!!$adotado->visitaum!!}}<br>
-            <strong>Data 2º visita:</strong> {{!!$adotado->visitadois!!}}<br>
-            <strong>Data 3º visita:</strong> {{!!$adotado->visitatres!!}}<br>
+            <strong>Data visita:</strong> {{!!$adotado->datavisita!!}}<br>
+
         </p>
     </div>
     <div class="card-footer">
+        @if ((Auth::check()) && (Auth::user()->isAdmin()))
             {{Form::open(['route' => ['adotados.destroy', $adotado->id],'method' => 'DELETE'])}}
             {{Form::submit('Excluir', ['class' => 'btn btn-outline-danger'])}}
+        @endif
             <a href="{{url('adotados/')}}" class="btn btn-outline-secondary">Voltar</a>
+        @if ((Auth::check()) && (Auth::user()->isAdmin()))
             {{Form::close()}}
+        @endif
     </div>
 </div>
 <br>
